+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/encodedautos/pushBlueF.java
@@ -1,8 +1,6 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.encodedautos;
 
-import com.acmerobotics.roadrunner.drive.Drive;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
@@ -10,9 +8,10 @@ import org.firstinspires.ftc.teamcode.main.DriveTrain;
 import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
-@Disabled
-@Autonomous (name="push Red Front")
-public class pushRedF extends OpMode{
+
+@Autonomous(name="PUSH Blue Front")
+public class pushBlueF extends OpMode {
+
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -23,14 +22,13 @@ public class pushRedF extends OpMode{
         encoded = new Encoded(hardwareMap, telemetry);
         driveTrain = new DriveTrain(hardwareMap, telemetry);
         visionProcessor = new FirstVisionProcessor();
-        visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"),visionProcessor);;
-
+        visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), visionProcessor);
     }
+
     @Override
     public void init_loop() {
         telemetry.addData("Identified", visionProcessor.getSelection());
     }
-
     @Override
     public void start() {
         visionPortal.stopStreaming();
@@ -38,34 +36,40 @@ public class pushRedF extends OpMode{
         switch (visionProcessor.getSelection()) {
             case LEFT:
                 encoded.closeClaw();
-                driveTrain.backward(33, 700);
-                driveTrain.turnLeft(20, 350);
-                driveTrain.backward(2, 700);
-                driveTrain.forward(3, 700);
+                encoded.armtoGroundAuto();
+                driveTrain.forward(29, 700);
+                driveTrain.turnLeft(20, 700);
+                driveTrain.backward(6, 500);
+                encoded.stopBot(1);
+                encoded.openBottomClaw();
+                driveTrain.backward(5, 700);
 
                 break;
 
             case NONE:
             case MIDDLE:
                 encoded.closeClaw();
-                driveTrain.backward(31, 500);
-                driveTrain.forward(8, 700);
-
+                encoded.armtoGroundAuto();
+                driveTrain.forward(18,700);
+                encoded.stopBot(1);
+                encoded.openBottomClaw();
+                driveTrain.backward(10, 700);
                 break;
 
             case RIGHT:
                 encoded.closeClaw();
-                driveTrain.backward(34, 700);
-                driveTrain.turnRight(20, 350);
-                driveTrain.backward(3, 500);
-                driveTrain.forward(4, 700);
-
+                encoded.armtoGroundAuto();
+                driveTrain.strafeRight(10,800);
+                driveTrain.forward(4,800);
+                encoded.stopBot(1);
+                encoded.openBottomClaw();
+                driveTrain.backward(2,800);
                 break;
 
         }
-
     }
     @Override
-    public void loop () {
+    public void loop() {
+
     }
 }
