+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/encodedautos/pushRedB.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.encodedautos;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -10,8 +10,8 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 @Disabled
-@Autonomous(name="Red Back")
-public class redBack extends OpMode {
+@Autonomous(name="PUSH Red Back")
+public class pushRedB extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -40,48 +40,18 @@ public class redBack extends OpMode {
                 driveTrain.backward(28,600);
                 driveTrain.turnLeft(20,700);
                 driveTrain.backward(3,600);
-                driveTrain.forward(3,600);
+                driveTrain.forward(2,600);
                 //purple scored
-                driveTrain.strafeLeft(8,700);
-                encoded.stopBot(0.5);
-                encoded.armScoreAuto();
-                encoded.backdropClawTilt();
-                driveTrain.forward(40,700);
-                encoded.openTopClaw();
-                encoded.stopBot(1);
-                driveTrain.backward(16,700);
-                //yellow pixel scored
-                encoded.closeClaw();
-                encoded.armtoGround();
-                driveTrain.strafeRight(40,700);
-                driveTrain.forward(10,700);
-                //robot parked
-
                 break;
 
             case NONE:
             case MIDDLE:
                 encoded.closeClaw();
-                driveTrain.backward(30,600);
+                driveTrain.backward(31,600);
                 encoded.retractTilt();
                 driveTrain.forward(4,700);
-                driveTrain.turnLeft(20.5,400);
-                encoded.stopBot(1);
+                driveTrain.turnLeft(19.5,400);
                 //purple scored
-                driveTrain.strafeLeft(3,400);
-                encoded.armScoreAuto();
-                encoded.backdropClawTilt();
-                driveTrain.forward(41,700);
-                encoded.openTopClaw();
-                encoded.stopBot(0.8);
-                //yellow scored
-                driveTrain.backward(10,700);
-                encoded.closeClaw();
-                encoded.armtoGround();
-                driveTrain.strafeRight(27,700);
-                driveTrain.forward(10,700);
-                //park
-
                 break;
 
             case RIGHT:
@@ -92,23 +62,6 @@ public class redBack extends OpMode {
                 encoded.stopBot(1);
                 driveTrain.backward(1,700);
                 driveTrain.forward(3,700);
-                //purple pixel scored
-                driveTrain.strafeLeft(16,700);
-                driveTrain.backward(3,600);
-                driveTrain.turnLeft(20,700);
-                encoded.stopBot(1);
-                encoded.armScoreAuto();
-                driveTrain.forward(15,700);
-                encoded.openTopClaw();
-                //yellow pixel scored
-                encoded.stopBot(1);
-                driveTrain.backward(12,700);
-                encoded.closeClaw();
-                encoded.armtoGround();
-                driveTrain.strafeRight(22,700);
-                driveTrain.forward(10,700);
-                //robot parked
-
                 break;
 
         }
