+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/encodedautos/pushRedF.java
@@ -1,6 +1,5 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.encodedautos;
 
-import com.acmerobotics.roadrunner.drive.Drive;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
@@ -11,8 +10,8 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 @Disabled
-@Autonomous(name="red Front")
-public class redFront extends OpMode {
+@Autonomous (name="push Red Front")
+public class pushRedF extends OpMode{
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -23,9 +22,9 @@ public class redFront extends OpMode {
         encoded = new Encoded(hardwareMap, telemetry);
         driveTrain = new DriveTrain(hardwareMap, telemetry);
         visionProcessor = new FirstVisionProcessor();
-        visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), visionProcessor);
-    }
+        visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"),visionProcessor);;
 
+    }
     @Override
     public void init_loop() {
         telemetry.addData("Identified", visionProcessor.getSelection());
@@ -37,31 +36,35 @@ public class redFront extends OpMode {
         telemetry.addData("Identified", visionProcessor.getSelection());
         switch (visionProcessor.getSelection()) {
             case LEFT:
-                driveTrain.forward(28,700);
-                driveTrain.turnLeft(23,700);
-                driveTrain.forward(4,700);
-                driveTrain.backward(2,700);
+                encoded.closeClaw();
+                driveTrain.backward(33, 700);
+                driveTrain.turnLeft(20, 350);
+                driveTrain.backward(2, 700);
+                driveTrain.forward(3, 700);
+
                 break;
 
             case NONE:
             case MIDDLE:
-                driveTrain.forward(30,900);
-                driveTrain.backward(3,900);
+                encoded.closeClaw();
+                driveTrain.backward(31, 500);
+                driveTrain.forward(8, 700);
+
                 break;
 
             case RIGHT:
-                driveTrain.forward(28,800);
-                driveTrain.turnRight(25,900);
-                driveTrain.forward(5,700);
-                driveTrain.backward(2,700);
+                encoded.closeClaw();
+                driveTrain.backward(34, 700);
+                driveTrain.turnRight(20, 350);
+                driveTrain.backward(3, 500);
+                driveTrain.forward(4, 700);
+
                 break;
 
         }
-    }
 
+    }
     @Override
-    public void loop() {
-
+    public void loop () {
     }
-
 }
