+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrautos/rrBlueFrontTRUSS.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.rrautos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -12,8 +12,9 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 @Disabled
-@Autonomous(name = "rrBlueFrontTRUSSyellowOnly")
-public class rrBlueFrontTRUSSyellowOnly extends OpMode {
+@Autonomous(name = "rrBlueFrontTRUSS")
+
+public class rrBlueFrontTRUSS extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -50,11 +51,23 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
+                        // if we have should
+                        .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
+                        .lineToConstantHeading(new Vector2d(42, 58.5)) // line up for truss
+                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // fly under truss
+                        .lineToLinearHeading(new Pose2d(-47, 35, Math.toRadians(180))) // orientate + to stack
+                        .lineToLinearHeading(new Pose2d(-55, 35, Math.toRadians(180))) // orientate + to stack
+                        .addTemporalMarker(0.5, ()-> {encoded.armtoPixelStack();})
+                        .addTemporalMarker(0.5, ()-> {encoded.closeClaw();})
+                        .lineToConstantHeading(new Vector2d(-50, 35))
+                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // orientate + line up for truss
+                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
+                        .splineToLinearHeading(new Pose2d(50, 35.5), Math.toRadians(0)) // to bd
+                        .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
+                        .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
-
-
                         .build();
 
                 break;
@@ -71,10 +84,23 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
+                        // if we have should
+                        .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
+                        .lineToConstantHeading(new Vector2d(42, 58.5)) // line up for truss
+                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // fly under truss
+                        .lineToLinearHeading(new Pose2d(-47, 35, Math.toRadians(180))) // orientate + to stack
+                        .lineToLinearHeading(new Pose2d(-55, 35, Math.toRadians(180))) // orientate + to stack
+                        .addTemporalMarker(0.5, ()-> {encoded.armtoPixelStack();})
+                        .addTemporalMarker(0.5, ()-> {encoded.closeClaw();})
+                        .lineToConstantHeading(new Vector2d(-50, 35))
+                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // orientate + line up for truss
+                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
+                        .splineToLinearHeading(new Pose2d(50, 35.5), Math.toRadians(0)) // to bd
+                        .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
+                        .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
-
                         .build();
 
 
@@ -92,6 +118,20 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
+                        // if we have should
+                        .lineToConstantHeading(new Vector2d(42, 28.5)) // back up from bd
+                        .lineToLinearHeading(new Pose2d(42, 58.5, Math.toRadians(180))) // line up for truss + orientate
+                        .lineToConstantHeading(new Vector2d(-34, 58.5)) // fly under truss
+                        .lineToConstantHeading(new Vector2d(-34, 23.5)) // drop down to middle stack y-cord
+                        .lineToConstantHeading(new Vector2d(-55, 23.5)) // to middle stack
+                        .addTemporalMarker(0.5, ()-> {encoded.armtoPixelStack();})
+                        .addTemporalMarker(0.5, ()-> {encoded.closeClaw();})
+                        .lineToConstantHeading(new Vector2d(-34, 23.5)) // back up
+                        .lineToConstantHeading(new Vector2d(-34, 58.5)) // line up for truss
+                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
+                        .splineToLinearHeading(new Pose2d(50, 28.5), Math.toRadians(0)) // to bd + angle
+                        .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
+                        .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .lineToConstantHeading(new Vector2d(42, 28.5)) // back up from bd
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
